Addressing review comments for VirtualServer refactor
