+++ f5_cccl/resource/ltm/profile/__init__.py
@@ -1,6 +1,4 @@
-"""Provides a class for managing BIG-IP Profile resources."""
-# coding=utf-8
-#
+#!/usr/bin/env python
 # Copyright 2017 F5 Networks Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,33 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+"""This module implements the F5 CCCL Profile class."""
 
-from f5_cccl.resource import Resource
 
-
-class Profile(Resource):
-    """Virtual Server class for managing configuration on BIG-IP."""
-
-    properties = dict(name=None,
-                      partition=None,
-                      context="all")
-
-    def __init__(self, name, partition, **properties):
-        """Create a Virtual server instance."""
-        super(Profile, self).__init__(name, partition)
-        self._data['context'] = properties.get('context', "all")
-
-    def __eq__(self, other):
-        if not isinstance(other, Profile):
-            return False
-
-        return super(Profile, self).__eq__(other)
-
-    def _uri_path(self, bigip):
-        """"""
-        raise NotImplementedError
-
-    def __repr__(self):
-        return 'Profile(%r, %r, context=%r)' % (self._data['name'],
-                                                self._data['partition'],
-                                                self._data['context'])
+from .profile import Profile  # noqa: F401
